+++ lib/app.dart
@@ -1,65 +1,16 @@
-import 'dart:io';
-
 import 'package:arquivolta/interfaces.dart';
 import 'package:arquivolta/logging.dart';
 import 'package:arquivolta/pages/install_page.dart';
+import 'package:arquivolta/platform/logger.dart';
 import 'package:arquivolta/services/job.dart';
-import 'package:arquivolta/services/util.dart';
 import 'package:beamer/beamer.dart';
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:get_it/get_it.dart';
-import 'package:logger/logger.dart';
-
-// ignore: implementation_imports
-import 'package:logger/src/outputs/file_output.dart';
-
-enum ApplicationMode { debug, production, test }
 
 typedef BeamerRouteList
     = Map<Pattern, dynamic Function(BuildContext, BeamState, Object?)>;
 
 typedef BeamerPageList = List<PageInfo>;
-
-Logger createLogger(ApplicationMode mode) {
-  if (mode == ApplicationMode.production && Platform.isWindows) {
-    final appData = getLocalAppDataPath();
-    final ourAppDataDir = Directory('$appData/Arquivolta')
-      ..createSync(recursive: true);
-
-    return Logger(
-      output: FileOutput(file: File('${ourAppDataDir.path}/log.txt')),
-      filter: ProductionFilter(),
-      printer: PrettyPrinter(
-        methodCount: 0,
-        errorMethodCount: 4,
-        excludeBox: {
-          Level.debug: true,
-          Level.info: true,
-          Level.verbose: true,
-        },
-        colors: false,
-        printEmojis: false, // NB: We add these in later
-      ),
-      level: Level.info,
-    );
-  }
-
-  // NB: filter: ProductionFilter is not a typo :facepalm:
-  return Logger(
-    filter: ProductionFilter(),
-    printer: PrettyPrinter(
-      methodCount: 0,
-      errorMethodCount: 4,
-      excludeBox: {
-        Level.debug: true,
-        Level.info: true,
-        Level.verbose: true,
-      },
-      colors: false,
-      printEmojis: false, // NB: We add these in later
-    ),
-  );
-}
 
 class App {
   static GetIt find = GetIt.instance;
@@ -71,15 +22,14 @@ class App {
       await locator.reset();
     }
 
-    final isTestMode =
-        Platform.isWindows && Platform.resolvedExecutable.contains('_tester');
+    final testMode = isTestMode();
     var isDebugMode = false;
 
     // NB: Assert statements are stripped from release mode. Clever!
     // ignore: prefer_asserts_with_message
     assert(isDebugMode = true);
 
-    final appMode = isTestMode
+    final appMode = testMode
         ? ApplicationMode.test
         : isDebugMode
             ? ApplicationMode.debug
