+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:arquivolta/app.dart';
 import 'package:fluent_ui/fluent_ui.dart';
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter_acrylic/flutter_acrylic.dart' as flutter_acrylic;
 import 'package:window_manager/window_manager.dart';
 
@@ -10,7 +10,7 @@ import 'package:window_manager/window_manager.dart';
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-  final desktopTasks = Platform.isWindows
+  final desktopTasks = !kIsWeb
       ? [
           flutter_acrylic.Window.initialize(),
           WindowManager.instance.ensureInitialized()
@@ -19,7 +19,7 @@ void main() async {
 
   await Future.wait([App.setupRegistration(), ...desktopTasks]);
 
-  if (Platform.isWindows) {
+  if (!kIsWeb) {
     unawaited(
       windowManager.waitUntilReadyToShow().then((_) async {
         await windowManager.setTitleBarStyle('hidden');
