+++ lib/platform/null/logger.dart
@@ -1,0 +1,10 @@
+import 'package:arquivolta/interfaces.dart';
+import 'package:logger/logger.dart';
+
+Logger createLogger(ApplicationMode mode) {
+  throw Exception('We should never get here');
+}
+
+bool isTestMode() {
+  return false;
+}
