+++ lib/platform/web/logger.dart
@@ -1,0 +1,24 @@
+import 'package:arquivolta/interfaces.dart';
+import 'package:logger/logger.dart';
+
+Logger createLogger(ApplicationMode mode) {
+  // NB: filter: ProductionFilter is not a typo :facepalm:
+  return Logger(
+    filter: ProductionFilter(),
+    printer: PrettyPrinter(
+      methodCount: 0,
+      errorMethodCount: 4,
+      excludeBox: {
+        Level.debug: true,
+        Level.info: true,
+        Level.verbose: true,
+      },
+      colors: false,
+      printEmojis: false, // NB: We add these in later
+    ),
+  );
+}
+
+bool isTestMode() {
+  return false;
+}
