+++ lib/platform/win32/install_arch.dart
@@ -2,10 +2,10 @@ import 'dart:io';
 import 'dart:ui' as ui;
 
 import 'package:arquivolta/logging.dart';
-import 'package:arquivolta/services/arch_to_rootfs.dart';
+import 'package:arquivolta/platform/win32/arch_to_rootfs.dart';
 import 'package:arquivolta/services/job.dart';
-import 'package:arquivolta/services/util.dart';
-import 'package:arquivolta/services/wsl.dart';
+import 'package:arquivolta/platform/win32/util.dart';
+import 'package:arquivolta/platform/win32/wsl.dart';
 import 'package:path/path.dart';
 import 'package:path_provider/path_provider.dart';
